# feat: improve performance on Node.js 16 by disabling class property definitions