--- a/src/always-encrypted/aead-aes-256-cbc-hmac-encryption-key.ts
+++ b/src/always-encrypted/aead-aes-256-cbc-hmac-encryption-key.ts
@@ -22,13 +22,13 @@ export const generateKeySalt = (
   `with encryption algorithm:${algorithmName} and key length:${keySize}`;
 
 export class AeadAes256CbcHmac256EncryptionKey extends SymmetricKey {
-  private readonly algorithmName: string;
-  private encryptionKeySaltFormat: string;
-  private macKeySaltFormat: string;
-  private ivKeySaltFormat: string;
-  private encryptionKey: SymmetricKey;
-  private macKey: SymmetricKey;
-  private ivKey: SymmetricKey;
+  declare private readonly algorithmName: string;
+  declare private encryptionKeySaltFormat: string;
+  declare private macKeySaltFormat: string;
+  declare private ivKeySaltFormat: string;
+  declare private encryptionKey: SymmetricKey;
+  declare private macKey: SymmetricKey;
+  declare private ivKey: SymmetricKey;
 
   constructor(rootKey: Buffer, algorithmName: string) {
     super(rootKey);