--- a/src/errors/timeout-error.ts
+++ b/src/errors/timeout-error.ts
@@ -1,5 +1,5 @@
 export default class TimeoutError extends Error {
-  code: string;
+  declare code: string;
 
   constructor() {
     super('The operation was aborted due to timeout');