--- a/src/outgoing-message-stream.ts
+++ b/src/outgoing-message-stream.ts
@@ -6,11 +6,11 @@ import Message from './message';
 import { Packet, HEADER_LENGTH } from './packet';
 
 class OutgoingMessageStream extends Duplex {
-  packetSize: number;
-  debug: Debug;
-  bl: any;
+  declare packetSize: number;
+  declare debug: Debug;
+  declare bl: any;
 
-  currentMessage: Message | undefined;
+  declare currentMessage: Message | undefined;
 
   constructor(debug: Debug, { packetSize }: { packetSize: number }) {
     super({ writableObjectMode: true });