--- a/src/rpcrequest-payload.ts
+++ b/src/rpcrequest-payload.ts
@@ -19,12 +19,12 @@ const STATUS = {
   s2.2.6.5
  */
 class RpcRequestPayload implements Iterable<Buffer> {
-  procedure: string | number;
-  parameters: Parameter[];
+  declare procedure: string | number;
+  declare parameters: Parameter[];
 
-  options: InternalConnectionOptions;
-  txnDescriptor: Buffer;
-  collation: Collation | undefined;
+  declare options: InternalConnectionOptions;
+  declare txnDescriptor: Buffer;
+  declare collation: Collation | undefined;
 
   constructor(procedure: string | number, parameters: Parameter[], txnDescriptor: Buffer, options: InternalConnectionOptions, collation: Collation | undefined) {
     this.procedure = procedure;