--- a/src/token/token-stream-parser.ts
+++ b/src/token/token-stream-parser.ts
@@ -7,9 +7,9 @@ import Message from '../message';
 import { TokenHandler } from './handler';
 
 export class Parser extends EventEmitter {
-  debug: Debug;
-  options: ParserOptions;
-  parser: Readable;
+  declare debug: Debug;
+  declare options: ParserOptions;
+  declare parser: Readable;
 
   constructor(message: Message, debug: Debug, handler: TokenHandler, options: ParserOptions) {
     super();