add declare for class level properties regard to #1579