@arthurschreiber 2023-10-06
https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#the-usedefineforclassfields-flag-and-the-declare-property-modifier has some more information around this, and the `declare` keyword on class properties. In a nutshell, `declare` just means "don't include this in the final JavaScript code", which is exactly what we want here. 👍