@arthurschreiber 2023-10-06
@MichaelSun90 Do you have the time to run a few of the benchmarks before / after this change and document the difference here in this issue? 🙇‍♂️